Career-tips page and pages added
