+++ src/pages/CareerTips.tsx
@@ -1,12 +1,10 @@
 import type React from "react"
 import { useState } from "react"
-import { useNavigate } from "react-router-dom"
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card"
 import { Input } from "@/components/ui/input"
 import { Button } from "@/components/ui/button"
-import { Badge } from "@/components/ui/badge"
-import { Search, Clock, Filter } from "lucide-react"
+import { Search, Filter } from "lucide-react"
 import { getCareerTipsFiltered, type CareerTip } from "@/lib/career-tips/career-tips"
+import TipCard from "@/components/CareerTips/TipCard"
 
 const categories = [
   { value: "all", label: "All Tips" },
@@ -20,27 +18,8 @@ const categories = [
 export default function CareerTips() {
   const [searchQuery, setSearchQuery] = useState(null)
   const [selectedCategory, setSelectedCategory] = useState<string>("all")
-  const navigate = useNavigate()
 
  const tips = getCareerTipsFiltered(selectedCategory as CareerTip["category"] as any, searchQuery);
-
-  const handleKeyPress = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter") {
-      // handleSearch()
-    }
-  }
-
-  const getCategoryColor = (category: CareerTip["category"]) => {
-    const colors = {
-      interview: "bg-blue-100 text-blue-800",
-      resume: "bg-green-100 text-green-800",
-      networking: "bg-purple-100 text-purple-800",
-      skills: "bg-orange-100 text-orange-800",
-      "career-growth": "bg-pink-100 text-pink-800",
-    }
-    return colors[category] || "bg-gray-100 text-gray-800"
-  }
-
 
   return (
     <div className="min-h-screen ">
@@ -61,7 +40,6 @@ export default function CareerTips() {
                 placeholder="Search career tips..."
                 value={searchQuery}
                 onChange={(e) => setSearchQuery(e.target.value)}
-                onKeyPress={handleKeyPress}
                 className="pl-10"
               />
             </div>
@@ -102,38 +80,7 @@ export default function CareerTips() {
         ) : (
           <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
             {tips.map((tip) => (
-              <Card
-                key={tip.id}
-                className="h-full hover:shadow-lg transition-shadow cursor-pointer bg-muted/30"
-                onClick={() => navigate(`/career-tips/${tip.slug}`)}
-              >
-                <CardHeader>
-                  <div className="flex items-start justify-between mb-2">
-                    <Badge className={getCategoryColor(tip.category)}>{tip.category.replace("-", " ")}</Badge>
-                    <div className="flex items-center text-sm text-gray-500">
-                      <Clock className="h-3 w-3 mr-1" />
-                      {tip.readTime} min
-                    </div>
-                  </div>
-                  <CardTitle className="text-lg leading-tight">{tip.title}</CardTitle>
-                  <CardDescription className="text-sm">{tip.excerpt}</CardDescription>
-                </CardHeader>
-                <CardContent>
-                  <div className="flex flex-wrap gap-1 mb-3">
-                    {tip.tags.slice(0, 3).map((tag) => (
-                      <Badge key={tag} variant="secondary" className="text-xs">
-                        {tag}
-                      </Badge>
-                    ))}
-                    {tip.tags.length > 3 && (
-                      <Badge variant="secondary" className="text-xs">
-                        +{tip.tags.length - 3}
-                      </Badge>
-                    )}
-                  </div>
-                  <p className="text-xs text-gray-500">Published {new Date(tip.publishedAt).toLocaleDateString()}</p>
-                </CardContent>
-              </Card>
+              <TipCard tip={tip} />
             ))}
           </div>
         )}
